Feature/move subclass in pantheon command (#1272)
* clean PantheonCommand class

- organize imports
- move `RpcApisConverter` and `RpcApisConversionException` to separate files

* spotlessApply

* register converters

* Revert "register converters"

This reverts commit aec47350792273615aeb2daab484799be928b47a.
